Refactor the wrapper to make the resolution of methods automated
Override __getattr__ and __call__ on a worker class so the resolution of
methods like api.linode.create() and api.linode.disk.create() are done
in a dynamic way.
Pass all final method calls through to a worker function that builds the
api kwargs and makes the final request.
Create a params file that defines the required variables for each API
method and deal with too many variables being passed in or methods that
only take optional arguments.